Algo/Trie: use better names and u8 instead of u16

+++ src/algorithms/trie.hpp
@@ -5,7 +5,7 @@
 
 #include "../utils/string_view.hpp"
 
-static constexpr uint16_t nil = ~uint16_t(0);
+static constexpr uint8_t nil = ~uint8_t(0);
 
 struct Trie {
 	struct Entry {
@@ -18,8 +18,8 @@ struct Trie {
 		Node(Node&&) = default;
 		Node(Node const&) = default;
 
-		uint16_t entry {nil};
-		std::array<uint16_t, 255> children;
+		uint8_t entry {nil};
+		std::array<uint8_t, 255> children;
 	};
 
 	std::vector<Node> nodes;
@@ -27,7 +27,7 @@ struct Trie {
 
 	Trie();
 
-	uint16_t go(uint16_t n, uint8_t c);
+	uint8_t go(uint8_t n, uint8_t c);
 	void insert(Entry);
 	bool has(string_view);
 	Entry longest_prefix_of(string_view);
